requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-cors-headers
django_compressor
django-libsass